refactor: componentize with ManualIssueExecutionUI
